requirements.txt
Flask
opencv-python
numpy
Pillow
sqlite3
Werkzeug